Add Spock unit tests to common submodule

+++ airline-approval-service/airline-approval-domain/airline-approval-domain-core/src/main/java/com/acroteq/ticketing/approval/service/domain/entity/order/OrderItem.java
@@ -9,6 +9,8 @@ import lombok.NonNull;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @Getter
 @ToString(callSuper = true)
 @SuperBuilder(toBuilder = true)
@@ -21,6 +23,6 @@ public class OrderItem extends MasterEntity<OrderItemId> {
 
   CashValue getAmount() {
     return flight.getPrice()
-                 .multiply(quantity);
+                 .multiply(BigDecimal.valueOf(quantity));
   }
 }
